Allow variable UMI Destinations in umi_external.rs
