+++ src/umi_external.rs
@@ -5,7 +5,7 @@ use std::path::PathBuf;
 
 use super::file_io;
 use crate::auxiliary::{threads_available, threads_per_task};
-use crate::read_editing::UMIDestination;
+use crate::read_editing::{umi_to_record_header,umi_to_record_seq,UMIDestination};
 use crate::umi_errors::RuntimeErrors;
 #[derive(Debug, Parser)]
 pub struct OptsExternal {
@@ -193,7 +193,11 @@ pub fn run(args: OptsExternal) -> Result<i32> {
         if r1_rec.id().eq(ru_rec.id()) {
             // Write to Output file
             let read_nr = if edit_nr { Some(1) } else { None };
-            let r1_rec = update_record(r1_rec, ru_rec.seq(), args.delim.as_ref(), read_nr)?;
+
+            let r1_rec = match args.target_position {
+                UMIDestination::Header => umi_to_record_header(r1_rec, ru_rec.seq(), args.delim.as_ref(), read_nr),
+                UMIDestination::Inline => umi_to_record_seq(r1_rec, ru_rec.seq(), args.delim.as_ref(), read_nr),
+            }?;
 
             write_output_r1.write_record(r1_rec)?;
         } else {
@@ -203,7 +207,11 @@ pub fn run(args: OptsExternal) -> Result<i32> {
         if r2_rec.id().eq(ru_rec.id()) {
             // Write to Output file
             let read_nr = if edit_nr { Some(2) } else { None };
-            let r2_rec = update_record(r2_rec, ru_rec.seq(), args.delim.as_ref(), read_nr)?;
+            
+            let r2_rec = match args.target_position {
+                UMIDestination::Header => umi_to_record_header(r2_rec, ru_rec.seq(), args.delim.as_ref(), read_nr),
+                UMIDestination::Inline => umi_to_record_seq(r2_rec, ru_rec.seq(), args.delim.as_ref(), read_nr),
+            }?;
 
             write_output_r2.write_record(r2_rec)?;
         } else {
@@ -214,26 +222,3 @@ pub fn run(args: OptsExternal) -> Result<i32> {
     Ok(counter)
 }
 
-// Updates the header and description of the reads accordingly
-fn update_record(
-    input: bio::io::fastq::Record,
-    umi: &[u8],
-    umi_sep: Option<&String>,
-    edit_nr: Option<u8>,
-) -> Result<bio::io::fastq::Record> {
-    let delim = umi_sep.as_ref().map(|s| s.as_str()).unwrap_or(":"); // the delimiter for the UMI
-    if let Some(number) = edit_nr {
-        let new_id = &[input.id(), delim, std::str::from_utf8(umi).unwrap()].concat();
-        let mut new_desc = String::from(input.desc().unwrap());
-        new_desc.replace_range(0..1, &number.to_string());
-        let desc: Option<&str> = Some(&new_desc);
-        let new_record =
-            bio::io::fastq::Record::with_attrs(new_id, desc, input.seq(), input.qual());
-        Ok(new_record)
-    } else {
-        let new_id = &[input.id(), delim, std::str::from_utf8(umi).unwrap()].concat();
-        let new_record =
-            bio::io::fastq::Record::with_attrs(new_id, input.desc(), input.seq(), input.qual());
-        Ok(new_record)
-    }
-}
